use only i32s

+++ src/graphics/draw/line.rs
@@ -1,10 +1,10 @@
 use crate::{Color, Context, Texture};
 
 pub fn draw(ctx: &mut Context, x0: i32, y0: i32, x1: i32, y1: i32, color: Color) {
-    let mut x0 = x0 as f32;
-    let mut x1 = x1 as f32;
-    let mut y0 = y0 as f32;
-    let mut y1 = y1 as f32;
+    let mut x0 = x0;
+    let mut x1 = x1;
+    let mut y0 = y0;
+    let mut y1 = y1;
 
     let dx = x0 - x1;
     let dy = y0 - y1;
@@ -18,7 +18,7 @@ pub fn draw(ctx: &mut Context, x0: i32, y0: i32, x1: i32, y1: i32, color: Color)
         let m = dy/dx;
         let mut y = y0;
 
-        for x in x0 as i32..=x1 as i32 {
+        for x in x0..=x1 {
             ctx.texture.set_pixel(x as usize, y as usize, color);
             y += m;
         }
@@ -32,7 +32,7 @@ pub fn draw(ctx: &mut Context, x0: i32, y0: i32, x1: i32, y1: i32, color: Color)
         let m = dx/dy;
         let mut x = x0;
 
-        for y in y0 as i32..=y1 as i32 {
+        for y in y0..=y1 {
             ctx.texture.set_pixel(x as usize, y as usize, color);
             x += m;
         }
